+++ fabric/src/main/java/com/diontryban/transparent/mixin/client/HumanoidArmorLayerMixinFabric.java
@@ -33,17 +33,12 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Redirect;
 
 @Mixin(HumanoidArmorLayer.class)
-public abstract class HumanoidArmorLayerMixin<T extends LivingEntity, M extends HumanoidModel<T>, A extends HumanoidModel<T>> extends RenderLayer<T, M> {
-    public HumanoidArmorLayerMixin(RenderLayerParent<T, M> renderer) {
+public abstract class HumanoidArmorLayerMixinFabric<T extends LivingEntity, M extends HumanoidModel<T>, A extends HumanoidModel<T>> extends RenderLayer<T, M> {
+    public HumanoidArmorLayerMixinFabric(RenderLayerParent<T, M> renderer) {
         super(renderer);
     }
 
-    // The Minecraft Development IntelliJ Plugin claims this is incorrect, but it is.
-    // On Fabric, there's only one renderModel method, but on NeoForge there are two, so it has to be specified.
-    @Redirect(
-            method = "renderModel(Lcom/mojang/blaze3d/vertex/PoseStack;Lnet/minecraft/client/renderer/MultiBufferSource;ILnet/minecraft/client/model/Model;ILnet/minecraft/resources/ResourceLocation;)V",
-            at = @At(value = "INVOKE", target = "Lnet/minecraft/client/renderer/RenderType;armorCutoutNoCull(Lnet/minecraft/resources/ResourceLocation;)Lnet/minecraft/client/renderer/RenderType;")
-    )
+    @Redirect(method = "renderModel", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/renderer/RenderType;armorCutoutNoCull(Lnet/minecraft/resources/ResourceLocation;)Lnet/minecraft/client/renderer/RenderType;"))
     public RenderType redirectArmorCutoutNoCull(ResourceLocation location) {
         return Transparent.CONFIG.playerArmor ?
                 TransparentRenderTypes.armorCutoutNoCull(location) :